// types.ts
export type Language = 'en' | 'tr' | 'hi';

export type View = 
  | 'dashboard' 
  | 'procurement-orders' 
  | 'procurement-prices' 
  | 'procurement-escalation' 
  | 'procurement-suppliers' 
  | 'procurement-bom' 
  | 'logistics-reservation'
  | 'logistics-transport'
  | 'warehouse-dashboard' 
  | 'warehouse-inefficiency'
  | 'customs-dashboard'
  | 'customs-import'
  | 'customs-export';